Preview Sheet set_url method issue fix, methods are added. The demo files are available in demo folder now.
