--- Sample_Documents/Notes.txt
@@ -1,2 +1,0 @@
-1349
-1009ad0ff781ba7b2e5130b9bd67324bea1e3512fad0182abb5baa7bab06c727c4f45ff768e9ab947536e0f005d3a7f3
