+++ configurations/AppConfiguration.json
