+++ demo/zohosheet/PreviewSpreadsheet.py
@@ -1,0 +1,48 @@
+from controllers.ClientSideException import ZOIException
+from controllers.Operations import PreviewSpreadsheet
+from controllers.RestClient import ZOIRestClient
+
+
+def configure():
+    """
+    Passing the entire configuration as a dictionary
+    """
+    try:
+        config_obj = ZOIRestClient.get_instance()
+        user_config = {
+            "apikey": "2ae438cf864488657cc9754a27daa480",
+            "writer_api_base_url": "https://api.office-integrator.com/writer/officeapi/",
+            "sheet_api_base_url": "https://api.office-integrator.com/sheet/officeapi/",
+            "show_api_base_url": "https://api.office-integrator.com/show/officeapi/"
+        }
+        config_obj.initialize(user_config)
+    except ZOIException as ex:
+        print(ex.status_code)
+        print(ex.error_code)
+        print(ex.error_message)
+        print(ex.error_details)
+        print(ex.error_content)
+
+
+def preview_spreadsheet():
+    try:
+        preview_sheet = PreviewSpreadsheet.get_instance()
+
+        preview_sheet.set_url("https://file-examples-com.github.io/uploads/2017/02/file_example_XLSX_5000.xlsx")
+        # preview_sheet.upload_document("document", "files/ZohoSheet.xlsx")
+
+        response = preview_sheet.preview_spreadsheet()
+        response_json = response.response_json
+        for key in response_json:
+            print("{0}: {1}".format(key, response_json[key]))
+    except ZOIException as ex:
+        print(ex.status_code)
+        print(ex.error_code)
+        print(ex.error_message)
+        print(ex.error_details)
+        print(ex.error_content)
+
+
+if __name__ == '__main__':
+    configure()
+    preview_spreadsheet()
