+++ demo/zohoshow/CreatePresentation.py
@@ -1,0 +1,49 @@
+from controllers.ClientSideException import ZOIException
+from controllers.Operations import CreatePresentation
+from controllers.RestClient import ZOIRestClient
+
+
+def configure():
+    """
+    Passing the entire configuration as a dictionary
+    """
+    try:
+        config_obj = ZOIRestClient.get_instance()
+        user_config = {
+            "apikey": "2ae438cf864488657cc9754a27daa480",
+            "writer_api_base_url": "https://api.office-integrator.com/writer/officeapi/",
+            "sheet_api_base_url": "https://api.office-integrator.com/sheet/officeapi/",
+            "show_api_base_url": "https://api.office-integrator.com/show/officeapi/"
+        }
+        config_obj.initialize(user_config)
+    except ZOIException as ex:
+        print(ex.status_code)
+        print(ex.error_code)
+        print(ex.error_message)
+        print(ex.error_details)
+        print(ex.error_content)
+
+
+def create_presentation():
+    try:
+        create_show = CreatePresentation.get_instance()
+
+        create_show.set_callback_settings("save_format", "pptx")
+        create_show.set_callback_settings("save_url", "https://domain.com/save.php")
+        create_show.set_callback_settings("context_info", "additional doc or user info")
+
+        response = create_show.create_presentation()
+        response_json = response.response_json
+        for key in response_json:
+            print("{0}: {1}".format(key, response_json[key]))
+    except ZOIException as ex:
+        print(ex.status_code)
+        print(ex.error_code)
+        print(ex.error_message)
+        print(ex.error_details)
+        print(ex.error_content)
+
+
+if __name__ == '__main__':
+    configure()
+    create_presentation()
