+++ demo/zohowriter/merge_api/GetFields.py
@@ -1,0 +1,48 @@
+from controllers.ClientSideException import ZOIException
+from controllers.Operations import GetFields
+from controllers.RestClient import ZOIRestClient
+
+
+def configure():
+    """
+    Passing the entire configuration as a dictionary
+    """
+    try:
+        config_obj = ZOIRestClient.get_instance()
+        user_config = {
+            "apikey": "2ae438cf864488657cc9754a27daa480",
+            "writer_api_base_url": "https://api.office-integrator.com/writer/officeapi/",
+            "sheet_api_base_url": "https://api.office-integrator.com/sheet/officeapi/",
+            "show_api_base_url": "https://api.office-integrator.com/show/officeapi/"
+        }
+        config_obj.initialize(user_config)
+    except ZOIException as ex:
+        print(ex.status_code)
+        print(ex.error_code)
+        print(ex.error_message)
+        print(ex.error_details)
+        print(ex.error_content)
+
+
+def get_fields():
+    try:
+        get_fields = GetFields.get_instance()
+
+        get_fields.upload_document("file_content", "../../../files/ZohoWriter_MergeTemplete.docx")
+        # get_fields.set_file_url("File URL Here")
+
+        response = get_fields.get_fields()
+        response_json = response.response_json
+        for key in response_json:
+            print("{0}: {1}".format(key, response_json[key]))
+    except ZOIException as ex:
+        print(ex.status_code)
+        print(ex.error_code)
+        print(ex.error_message)
+        print(ex.error_details)
+        print(ex.error_content)
+
+
+if __name__ == '__main__':
+    configure()
+    get_fields()
